erased layout and moved footer to components folder
