+++ client/src/App.jsx
@@ -1,5 +1,4 @@
 import { Route, Switch } from "wouter";
-import { Layout } from "./components";
 import {
   ActiveEmps,
   EditEmp,
@@ -11,16 +10,14 @@ import {
 
 function App() {
   return (
-    <Layout>
-      <Switch>
-        <Route path="/" component={ActiveEmps} />
-        <Route path="/edit/:id" component={EditEmp} />
-        <Route path="/details/:id" component={EmpDetails} />
-        <Route path="/logs" component={EmpLogs} />
-        <Route path="/register" component={RegisterEmp} />
-        <Route path="/hours" component={WorkedHours} />
-      </Switch>
-    </Layout>
+    <Switch>
+      <Route path="/" component={ActiveEmps} />
+      <Route path="/edit/:id" component={EditEmp} />
+      <Route path="/details/:id" component={EmpDetails} />
+      <Route path="/logs" component={EmpLogs} />
+      <Route path="/register" component={RegisterEmp} />
+      <Route path="/hours" component={WorkedHours} />
+    </Switch>
   );
 }
