--- client/src/components/Layout/index.jsx
@@ -1,13 +1,0 @@
-import Footer from "./Footer/Index"
-import './styles.css'
-
-const index = ({children}) => {
-  return (
-    <>
-      {children}
-      <Footer />
-    </>
-  )
-}
-
-export default index
