--- client/src/components/Layout/styles.css
@@ -1,5 +1,0 @@
-.Footer {
-    position: fixed;
-    bottom: 0;
-    width: 100vw;
-}
