+++ client/src/components/index.js
@@ -1,4 +1,4 @@
 export { default as Header } from "./Header";
-export { default as Layout } from "./Layout";
 export { default as Form } from "./Form";
 export { default as EditPhoto } from "./EditPhoto";
+export { default as Footer} from "./Footer";
